directory restructure and tools factory directory setup
