+++ tools/buck_visualizer/news_overlay_plot.py
@@ -6,7 +6,7 @@ from datetime import datetime
 import plotly.graph_objects as go
 import pandas as pd
 
-from data_provider_viz import DataVisualizationDownloader, fix_imports
+from tools.buck_visualizer.data_provider_viz import DataVisualizationDownloader, fix_imports
 
 fix_imports()
